Delete the player's stores, rename player/ to audio/, and guard the line (ADR-0083)
3,471 lines. `playerStore`, `queueStore` (1,016 lines on its own), `playbackStore` and their tests,
plus `useScrobbling` and `usePreviewAudio`, which had zero references anywhere.

Two more consumers had to be unpinned first, and both were doing nothing:

  PlayIndicator          `isLoadingAudio` — every other input was already a prop
  useTrackContextMenu    `addToQueue` and `setQueue`, on a page that owns no queue

The context menu is the better example. "Add to queue" and "Play next" render on `/embed`, where
the store is not mounted, so both wrote into a queue nothing plays from. Two menu items that did
nothing, with no error. They are now supplied by a surface that has a queue, or absent — which is
the honest rendering.

**`player/` becomes `audio/`.** What survives is an audio-engine seam, not a player: the engine
registry, the analysis buffers, the interceptor, and the queue-source type an API call still uses.
Nothing left in the tree is called "player" (ADR-0083 point 4). The web build caught the one import
the rewrite missed — `@familiar/frontend/src/player/...` in `embed.tsx`, package-qualified rather
than relative.

**A dependency-cruiser rule replaces `check-audio-guardrails.mjs`**, which asserted the same
property by `readFileSync`-ing two files it named by path — so it would have *crashed* when they
were deleted rather than reporting anything. The new rule allows the surfaces' own stores by name
(columns, visualizer choice, plugin refusals) and forbids the rest, because the failure it exists
to catch is invisible: a selector against an unmounted store returns nothing, renders nothing
wrong, and quietly pins the graph behind it into a bundle that cannot play anything.

Verified the rule is not vacuous: a planted `stores/homeStore` import in a Discovery component is
caught, and removing it returns 0 errors across 274 modules and 864 dependencies.

362 frontend tests (down from 489: the deleted stores took their own tests with them). Web and
visualizer builds both green.
