+++ packages/frontend/src/audio/__tests__/analysisMetrics.test.ts
