--- packages/frontend/src/stores/playerStore.ts
@@ -1,2 +1,0 @@
-// DEPRECATED: import from '../player' instead
-export * from '../player/playerStore';
